# dataing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dataing",
    "description": "Autonomous Data Quality Investigation",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/permissions/": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "List Permissions",
        "description": "List all permission grants in the organization.",
        "operationId": "list_permissions_api_v1_permissions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "permissions"
        ],
        "summary": "Create Permission",
        "description": "Create a new permission grant.\n\nRequires admin scope.",
        "operationId": "create_permission_api_v1_permissions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGrantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionGrantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/{grant_id}": {
      "delete": {
        "tags": [
          "permissions"
        ],
        "summary": "Delete Permission",
        "description": "Delete a permission grant.\n\nRequires admin scope.",
        "operationId": "delete_permission_api_v1_permissions__grant_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/investigations/{investigation_id}/permissions/": {
      "get": {
        "tags": [
          "investigation-permissions"
        ],
        "summary": "Get Investigation Permissions",
        "description": "Get all permissions for an investigation.",
        "operationId": "get_investigation_permissions_api_v1_investigations__investigation_id__permissions__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionGrantResponse"
                  },
                  "title": "Response Get Investigation Permissions Api V1 Investigations  Investigation Id  Permissions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionGrantCreate": {
        "properties": {
          "grantee_type": {
            "type": "string",
            "enum": [
              "user",
              "team"
            ],
            "title": "Grantee Type"
          },
          "grantee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Grantee Id"
          },
          "access_type": {
            "type": "string",
            "enum": [
              "resource",
              "tag",
              "datasource"
            ],
            "title": "Access Type"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type",
            "default": "investigation"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "tag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Id"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "permission": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "admin"
            ],
            "title": "Permission"
          }
        },
        "type": "object",
        "required": [
          "grantee_type",
          "grantee_id",
          "access_type",
          "permission"
        ],
        "title": "PermissionGrantCreate",
        "description": "Permission grant creation request."
      },
      "PermissionGrantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "grantee_type": {
            "type": "string",
            "title": "Grantee Type"
          },
          "grantee_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grantee Id"
          },
          "access_type": {
            "type": "string",
            "title": "Access Type"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "tag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Id"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "permission": {
            "type": "string",
            "title": "Permission"
          }
        },
        "type": "object",
        "required": [
          "id",
          "grantee_type",
          "grantee_id",
          "access_type",
          "resource_type",
          "resource_id",
          "tag_id",
          "data_source_id",
          "permission"
        ],
        "title": "PermissionGrantResponse",
        "description": "Permission grant response."
      },
      "PermissionListResponse": {
        "properties": {
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionGrantResponse"
            },
            "type": "array",
            "title": "Permissions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "permissions",
          "total"
        ],
        "title": "PermissionListResponse",
        "description": "Response for listing permissions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}